request handling moved to separate module
