+++ src/main/kotlin/com/example/plugins/Routing.kt
@@ -1,20 +1,11 @@
 package com.example.plugins
 
-import com.example.api.requests.*
-import com.example.api.responses.AuthResponse
+import com.example.api.*
 import com.example.database.hasConnection
-import com.example.repository.ArticlesRepository
-import com.example.repository.UsersRepository
-import com.example.repository.impl.DatabaseArticlesRepository
-import com.example.repository.impl.DatabaseUsersRepository
-import com.example.repository.impl.DefaultArticlesRepository
-import com.example.repository.impl.DefaultUsersRepository
-import com.example.services.impl.DefaultJWTService
-import io.ktor.http.*
+import com.example.repository.*
+import com.example.repository.impl.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
-import io.ktor.server.request.*
-import io.ktor.server.response.*
 import io.ktor.server.routing.*
 
 
@@ -30,92 +21,35 @@ fun Application.configureRouting() {
         users = DefaultUsersRepository()
     }
     values.createArticle("example", "unknown")
-    val tokenService = DefaultJWTService()
 
     routing {
         post("/registration") {
-            val request = call.receive<RegistrationRequest>()
-            val result = users.createUser(request.login, request.password, request.name)
-            result ?: call.respond(HttpStatusCode.Conflict)
-            call.respond(HttpStatusCode.OK)
+            handleRegistration(call, values, users)
         }
 
         post("/login") {
-            val request = call.receive<LoginRequest>()
-            if (!users.hasLogin(request.login)) {
-                call.respond(HttpStatusCode.Unauthorized)
-            }
-            val correctPassword = users.getByLogin(request.login)?.password!!
-            if (correctPassword == request.password) {
-                val token = tokenService.createAccessToken(request.login)
-                call.respond(AuthResponse(token))
-            } else {
-                call.respond(HttpStatusCode.Unauthorized)
-            }
-
+            handleLogin(call, values, users)
         }
         get("/") {
-            call.respondText("Hello, world!")
+            handleRoot(call, values, users)
         }
 
         get("/articles/all") {
-            val comments = values.getAll()
-            call.respond(comments)
+            handleGetAll(call, values, users)
         }
 
         get("/articles/{id}") {
-            var id = call.parameters["id"]?.toInt()
-            if (id == null) {
-                call.respond(HttpStatusCode.NotFound)
-            }
-            id = id!!
-            if (values.hasId(id)) {
-                call.respond(values.getById(id)!!)
-            }
-            else {
-                call.respond(HttpStatusCode.NotFound)
-            }
+            handleGet(call, values, users)
         }
         authenticate("auth-jwt") {
             put("/articles/create") {
-                val request = call.receive<CreateArticleRequest>()
-                if (request.articleText.length > 500) {
-                    call.respond(HttpStatusCode.PreconditionFailed)
-                }
-                val username = tokenService.extractUsername(request.token)
-                username ?: call.respond(HttpStatusCode.BadRequest)
-                val createdArticle = values.createArticle(request.articleText, username!!)
-                call.respond(createdArticle)
+                handleCreate(call, values, users)
             }
             patch("/articles/update") {
-                val request = call.receive<UpdateArticleRequest>()
-                if (!values.hasId(request.id)) {
-                    call.respond(HttpStatusCode.NotFound)
-                }
-                val username = tokenService.extractUsername(request.token)
-                username ?: call.respond(HttpStatusCode.Unauthorized)
-                if (username != values.getById(request.id)?.author) {
-                    call.respond(HttpStatusCode.Forbidden)
-                }
-                if (request.articleText.length > 500) {
-                    call.respond(HttpStatusCode.PreconditionFailed)
-                }
-
-                values.updateById(request.id, request.articleText)
-                call.respond(values.getById(request.id)!!)
+                handleUpdate(call, values, users)
             }
             delete("/articles/delete") {
-                val request = call.receive<DeleteArticleRequest>()
-                if (!values.hasId(request.id)) {
-                    call.respond(HttpStatusCode.NotFound)
-                }
-                val username = tokenService.extractUsername(request.token)
-                username ?: call.respond(HttpStatusCode.Unauthorized)
-                if (username != values.getById(request.id)?.author) {
-                    call.respond(HttpStatusCode.Forbidden)
-                }
-                values.dropId(request.id)
-                call.respond(HttpStatusCode.OK)
+                handleDelete(call, values, users)
             }
         }
     }
